+++ dashboard/src/components/TrackItemOptions.tsx
@@ -4,7 +4,7 @@ import { Button, Form, Modal, Stack } from "react-bootstrap";
 import { useParams } from "react-router-dom";
 import { useFetchServerQuery } from "store/api";
 import { setItemDeathsChannel, setItemKillsChannel } from "store/track";
-import { ITrackItem, TRACK_TYPE } from "types";
+import { ITrackItem, TRACK_TYPE } from "types/track";
 import ChannelInput from "./ChannelInput";
 
 interface ITrackItemOptionsProps {
