+++ dashboard/src/components/subscriptions/SubscriptionAdd.tsx
@@ -3,7 +3,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { useEffect, useState } from "react";
 import { Button, Modal } from "react-bootstrap";
 import { useCreateAdminSubscriptionMutation } from "store/api/admin";
-import { ISubscription } from "types";
+import { ISubscription } from "types/subscription";
 import SubscriptionForm from "./SubscriptionForm";
 
 const SubscriptionAdd = () => {
