+++ dashboard/src/types/limits.ts
@@ -1,0 +1,5 @@
+export interface Limits {
+  players: number;
+  guilds: number;
+  alliances: number;
+}
