+++ dashboard/src/types/user.ts
@@ -1,0 +1,7 @@
+export interface User {
+  id: string;
+  username: string;
+  avatar: string;
+  locale: string;
+  admin: boolean;
+}
